feat: #87 LocationClient 구현 및 DetailFeature 연결
