+++ Projects/Client/Location/Sources/Client/LocationClient.swift
@@ -11,7 +11,7 @@ import Shared
 
 @DependencyClient
 public struct LocationClient: Sendable {
-  public var reqeustUserLocation: @Sendable () async throws -> Coordinate?
+  public var requestUserLocation: @Sendable () async throws -> Coordinate?
 }
 
 public extension DependencyValues {
