+++ Projects/Client/Location/Sources/Entity/entityDummy.swift
