+++ Projects/Feature/Map/MapFeature/Sources/MapView/LocationFeature.swift
@@ -13,11 +13,13 @@ import ComposableArchitecture
 import MapFeatureInterface
 import FlowerSpotClient
 import CacheClient
+import LocationClient
 
 extension MapFeature {
   struct LocationFeature: Reducer {
     typealias Action = LocationAction
     
+    @Dependency(\.locationClient) var locationClient
     @Dependency(\.flowerSpotClient) var flowerSpotClient
     @Dependency(\.cache) var cache
 
@@ -74,7 +76,7 @@ extension MapFeature {
 extension MapFeature.LocationFeature {
   private func moveUserLocation(isCurrentButtonTap: Bool) -> Effect<Action> {
     return .run { send in
-      if let location = await LocationService.shared.requestUserLocation() {
+      if let location = try await locationClient.requestUserLocation() {
         await send(.saveUserLocation(location))
         await send(.moveLocation(location))
       } else {
