+++ Projects/Feature/Map/Project.swift
@@ -14,6 +14,7 @@ let project = Project.buildFeature(
   interfaceDependencies: [
     .Client.FlowerSpot,
     .Client.Blooming,
+    .Client.Location,
     .SPM.NMap
   ]
 )
